+++ fluent_contents/extensions/__init__.py
@@ -8,6 +8,7 @@ The API uses a registration system.
 While plugins can be easily detected via ``__subclasses__()``, the register approach is less magic and more explicit.
 Having to do an explicit register ensures future compatibility with other API's like reversion.
 """
+from .model_fields import PluginUrlField, PluginFileField, PluginImageField
 from .pluginbase import PluginContext, ContentPlugin
 from .pluginpool import plugin_pool, PluginNotFound, PluginAlreadyRegistered
 from fluent_contents.forms import ContentItemForm   # Expose over here now, still leave at old location.
@@ -17,6 +18,7 @@ from fluent_contents.models import ContentItem      # Allow plugins to pick ever
 __all__ = (
     'PluginContext', 'ContentPlugin',
     'ContentItem',
+    'PluginUrlField', 'PluginFileField', 'PluginImageField',
     'ContentItemForm',
     'plugin_pool',
     'PluginNotFound', 'PluginAlreadyRegistered',